requirements.txt
environs[django]==9.5.*
django==4.1.*
psycopg2-binary==2.9.5
django-debug-toolbar
flake8
flake8-isort
djangorestframework==3.14.0
django-filter
drf_yasg
django-phonenumber-field[phonenumbers]==7.0.*